% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Harmonizer.R
\name{harmonize}
\alias{harmonize}
\title{Harmonize the various databases}
\usage{
harmonize(df, col_map, race_encoding, date_format, name_delim, null_name,
  null_race, null_age)
}
\arguments{
\item{df}{Dataframe with scraped data}

\item{col_map}{a named list of columns that appear in df associated with
their canonical column names. The final data frame must possess a subset of
columns with specified names and data types. Columns which are not listed
will be included in the final data frame unchanged.}

\item{race_encoding}{a named list which maps the races in df to something standard}

\item{date_format}{a regex string representing the way the date is formatted in df}

\item{name_delim}{a regex string representing the delimiter
between a person name and alias}

\item{null_name}{the symbol used to signify null name}

\item{null_race}{the symbol used to signify null race}

\item{null_age}{the symbol used to signify null age}
}
\description{
Harmonize the various databases
}
