% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasterScraper.R
\name{scrape_all_data}
\alias{scrape_all_data}
\title{---
title: "Scraping functions for various online repositories"
author: "Martina Morris"
author: "Vaughn Johnson"
date: '`r format(Sys.Date(), "%Y-%m-%d")`'
---
A compliation of functions which scrape data from Mapping Police Violence,
Killed By Police, Fatal Encounters, and the Census.}
\usage{
scrape_all_data()
}
\value{
Void. Adds all scraped data to /ScrapedFiles/ dir
}
\description{
Always assume we're stating in the ffsg top level directory
Get data from Census, KBP, FE, and MPV
}
