% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeFEData.R
\name{scrape_FE_data}
\alias{scrape_FE_data}
\title{Scrape the Census Bureau's website for population data}
\usage{
scrape_FE_data(url, save_file)
}
\arguments{
\item{url}{URL for Fatal Encounters spreadsheet}

\item{save_file}{filename to be saved in /ScrapedFiles/}
}
\value{
Void. Saves the data to `save_file`.
}
\description{
Scrape the Census Bureau's website for population data
}
