% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeKBPData.R
\name{scrape_KBP_data}
\alias{scrape_KBP_data}
\title{Scrape and Parse Killed By Police's online tables}
\usage{
scrape_KBP_data(save_file)
}
\arguments{
\item{save_file}{filename to be saved in /ScrapedFiles/}
}
\value{
Void. Saves the data to `save_file`.
}
\description{
Scrape and Parse Killed By Police's online tables
}
