% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeMPVData.R
\name{scrape_MPV_data}
\alias{scrape_MPV_data}
\title{Scrape the MPV database}
\usage{
scrape_MPV_data(url, save_file)
}
\arguments{
\item{url}{URL for MPV spreadsheet}

\item{save_file}{filename to be saved in /ScrapedFiles/}
}
\value{
Void. Saves the data to `save_file`.
}
\description{
Scrape the MPV database
}
