% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePopData.R
\name{scrape_population_data}
\alias{scrape_population_data}
\title{Scrape the Census Bureau's website for population data}
\usage{
scrape_population_data(state_urls, county_urls, save_file)
}
\arguments{
\item{state_urls}{A vector of the URLS for state and regional data}

\item{county_urls}{A vector of the URLS for county data}
}
\value{
Void. Saves the data to `save_file`.
}
\description{
Scrape the Census Bureau's website for population data
}
